feat: rename project from Claudeee to CCDash
Major project rebranding and renaming changes:

## Package and Binary Changes
- Update npm package name: claudeee → ccdash
- Rename main CLI script: bin/claudeee.js → bin/ccdash.js
- Update binary names: claudeee-server → ccdash-server
- Update Go module name: claudeee-backend → ccdash-backend

## Configuration and Environment
- Update environment variable: CLAUDEEE_DB_PATH → CCDASH_DB_PATH
- Update database filename: claudeee.db → ccdash.db
- Update database directory: ~/.claudeee → ~/.ccdash
- Update localStorage keys: claudeee-* → ccdash-*

## User Interface and Branding
- Update project title: "Claudeee" → "CCDash"
- Update CLI logo with new CCDASH ASCII art
- Update all help text and usage examples
- Update frontend metadata and page titles
- Update i18n dictionary entries

## Documentation and Configuration
- Update all README files and documentation
- Update GitHub Actions workflows
- Update Docker configurations and container names
- Update Makefile project name
- Update all command-line tools and scripts

## Repository and URLs
- Update GitHub repository URLs
- Update npm package repository references
- Update issue tracking URLs

This comprehensive rename maintains full backward compatibility
while establishing the new CCDash brand identity.
