+++ README.md
@@ -1,10 +1,10 @@
-# Claudeee
+# CCDash
 
-Claudeee is a web application that monitors the execution status of Claude Code and performs task scheduling.
+CCDash is a web application that monitors the execution status of Claude Code and performs task scheduling.
 
 ## 📋 Supported Platforms
 
-Claudeee supports the following platforms:
+CCDash supports the following platforms:
 
 | Platform | Architecture | Status |
 |----------|-------------|---------|
@@ -19,14 +19,14 @@ Claudeee supports the following platforms:
 
 ```bash
 # Run instantly using NPX
-npx claudeee
+npx ccdash
 
 # Or, global installation
-npm install -g claudeee
-claudeee
+npm install -g ccdash
+ccdash
 
 # Start with custom ports
-claudeee --backend-port 8081 --frontend-port 3001
+ccdash --backend-port 8081 --frontend-port 3001
 ```
 
 After installation, access the following URLs in your browser:
@@ -36,7 +36,7 @@ After installation, access the following URLs in your browser:
 
 ## Overview
 
-Claudeee is a tool that visualizes Claude Code usage and helps with efficient task management. It supports monitoring token usage, managing sessions, and will support automatic task scheduling in the future.
+CCDash is a tool that visualizes Claude Code usage and helps with efficient task management. It supports monitoring token usage, managing sessions, and will support automatic task scheduling in the future.
 
 ## Features
 
@@ -87,7 +87,7 @@ Claudeee is a tool that visualizes Claude Code usage and helps with efficient ta
 ## Architecture
 
 ```
-claudeee/
+ccdash/
 ├── backend/          # Go/Gin Backend
 │   ├── cmd/server/   # Application entry point
 │   ├── internal/     # Internal packages
@@ -141,7 +141,7 @@ claudeee/
 
     ```bash
     git clone <repository-url>
-    cd claudeee
+    cd ccdash
     ```
 
 2.  **Backend setup**
@@ -162,29 +162,29 @@ claudeee/
 
 ## Commands
 
-Claudeee offers the following commands:
+CCDash offers the following commands:
 
 ```bash
 # Start the application (production mode)
-npx claudeee
-npx claudeee start
+npx ccdash
+npx ccdash start
 
 # Start with custom ports
-npx claudeee --backend-port 8081 --frontend-port 3001
-npx claudeee -bp 8081 -fp 3001
+npx ccdash --backend-port 8081 --frontend-port 3001
+npx ccdash -bp 8081 -fp 3001
 
 # Start in development mode (with hot-reloading)
-npx claudeee dev
-npx claudeee dev --backend-port 8081
+npx ccdash dev
+npx ccdash dev --backend-port 8081
 
 # Build the application
-npx claudeee build
+npx ccdash build
 
 # Display help
-npx claudeee help
+npx ccdash help
 
 # Display version
-npx claudeee version
+npx ccdash version
 ```
 
 ### Command Line Options
@@ -203,14 +203,14 @@ npx claudeee version
 ### Starting in Production Environment
 
 ```bash
-npx claudeee
+npx ccdash
 ```
 
 ### Starting in Development Environment
 
 ```bash
 # Start in development mode (recommended)
-npx claudeee dev
+npx ccdash dev
 
 # Or manually start each service
 ```
@@ -289,7 +289,7 @@ Example token usage response:
 #### Backend
 
   - `GIN_MODE`: Gin operation mode (development/release)
-  - `DB_PATH`: Path to the database file (default: `~/.claudeee/claudeee.db`)
+  - `DB_PATH`: Path to the database file (default: `~/.ccdash/ccdash.db`)
 
 #### Frontend
 
@@ -297,7 +297,7 @@ Example token usage response:
 
 ### Claude Code Configuration
 
-Claudeee parses JSONL log files generated by Claude Code.
+CCDash parses JSONL log files generated by Claude Code.
 Log file location: `~/.claude/projects/{project-name}/{session-id}.jsonl`
 
 ## Troubleshooting
@@ -307,7 +307,7 @@ Log file location: `~/.claude/projects/{project-name}/{session-id}.jsonl`
 1.  **Database Lock Error**
 
     ```bash
-    rm -f ~/.claudeee/claudeee.db*
+    rm -f ~/.ccdash/ccdash.db*
     ```
 
 2.  **Frontend Dependency Error**
