+++ backend/cmd/sync-reset/main.go
@@ -23,7 +23,7 @@ func main() {
 		os.Exit(1)
 	}
 
-	dbPath := filepath.Join(homeDir, ".claudeee", "claudeee.db")
+	dbPath := filepath.Join(homeDir, ".ccdash", "ccdash.db")
 	if _, err := os.Stat(dbPath); os.IsNotExist(err) {
 		fmt.Println("Database does not exist. No sync states to reset.")
 		return
