+++ backend/go.mod
@@ -1,4 +1,4 @@
-module claudeee-backend
+module ccdash-backend
 
 go 1.24.4
