+++ backend/internal/services/diff_sync_service_test.go
@@ -6,7 +6,7 @@ import (
 	"testing"
 	"time"
 
-	"claudeee-backend/internal/models"
+	"ccdash-backend/internal/models"
 	_ "github.com/marcboeker/go-duckdb"
 )
