+++ backend/internal/services/session_service.go
@@ -7,7 +7,7 @@ import (
 	"strings"
 	"time"
 	
-	"claudeee-backend/internal/models"
+	"ccdash-backend/internal/models"
 )
 
 type SessionService struct {
