+++ frontend/package.json
@@ -1,5 +1,5 @@
 {
-  "name": "claudeee",
+  "name": "ccdash",
   "version": "0.4.0",
   "private": true,
   "scripts": {
