+++ package.json
@@ -1,15 +1,15 @@
 {
-  "name": "claudeee",
+  "name": "ccdash",
   "version": "0.4.1",
-  "description": "Claude Code Monitoring - A comprehensive monitoring for Claude Code",
-  "main": "bin/claudeee.js",
+  "description": "Claude Code Dashboard - A comprehensive monitoring dashboard for Claude Code",
+  "main": "bin/ccdash.js",
   "bin": {
-    "claudeee": "./bin/claudeee.js"
+    "ccdash": "./bin/ccdash.js"
   },
   "scripts": {
-    "start": "node bin/claudeee.js",
+    "start": "node bin/ccdash.js",
     "build": "npm run build:frontend && npm run build:backend",
-    "build:backend": "if [ -d \"backend\" ]; then cd backend && go build -o ../bin/claudeee-server cmd/server/main.go; else echo \"Backend directory not found, skipping backend build (likely CI environment)\"; fi",
+    "build:backend": "if [ -d \"backend\" ]; then cd backend && go build -o ../bin/ccdash-server cmd/server/main.go; else echo \"Backend directory not found, skipping backend build (likely CI environment)\"; fi",
     "build:frontend": "if [ -d \"frontend\" ] && [ \"$CI\" != \"true\" ]; then cd frontend && (npm ci || npm install) && npx next build; else echo \"Skipping frontend build (CI environment or frontend directory not found)\"; fi",
     "dev": "npm run dev:backend & npm run dev:frontend",
     "dev:backend": "cd backend && go run cmd/server/main.go",
@@ -28,16 +28,16 @@
     "development-tools",
     "cli"
   ],
-  "author": "Claude Code Team",
+  "author": "CCDash team",
   "license": "MIT",
   "repository": {
     "type": "git",
-    "url": "git+https://github.com/claudeee/claudeee.git"
+    "url": "git+https://github.com/satoshi03/ccdash.git"
   },
   "bugs": {
-    "url": "https://github.com/claudeee/claudeee/issues"
+    "url": "https://github.com/satoshi03/ccdash/issues"
   },
-  "homepage": "https://github.com/claudeee/claudeee#readme",
+  "homepage": "https://github.com/satoshi03/ccdash#readme",
   "engines": {
     "node": ">=18.0.0",
     "npm": ">=8.0.0"
